Refactor and improve calendar overview and navigation UI
- Update ContentView and CalendarOverviewSheet with more consistent and modern design
- Enhance calendar item layout with improved typography and spacing
- Add tap gestures and context menu for calendar editing and deletion
- Separate CreateCalendarView and EditCalendarView into dedicated files
- Improve navigation title display modes across multiple views
